+++ src/librpbase/config/Config.hpp
@@ -47,10 +47,10 @@ public:
 public:
 	/** Image types **/
 
-	// Image type priority data.
+	// Image type priority data
 	struct ImgTypePrio_t {
-		const uint8_t *imgTypes;	// Image types.
-		uint32_t length;		// Length of imgTypes array.
+		const uint8_t *imgTypes;	// Image types
+		size_t length;			// Length of imgTypes array
 	};
 
 	// TODO: Function to get image type priority for a specified class.
